cambios varios

+++ ProjectsControl/Controllers/OffersController.cs
@@ -6,10 +6,9 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using ProjectsControl.Models;
-using Microsoft.AspNetCore.Authorization;
+
 namespace ProjectsControl.Controllers
 {
-    [Authorize]
     public class OffersController : Controller
     {
         private readonly DBProjectContext _context;
@@ -20,7 +19,6 @@ namespace ProjectsControl.Controllers
         }
 
         // GET: Offers
-        [AllowAnonymous]
         public async Task<IActionResult> Index()
         {
             var dBProjectContext = _context.Offers.Include(o => o.Customer);
@@ -30,7 +28,7 @@ namespace ProjectsControl.Controllers
         // GET: Offers/Details/5
         public async Task<IActionResult> Details(string id)
         {
-            if (id == null)
+            if (id == null || _context.Offers == null)
             {
                 return NotFound();
             }
@@ -47,21 +45,18 @@ namespace ProjectsControl.Controllers
         }
 
         // GET: Offers/Create
-        [Authorize(Roles = "Admin,Editor,Sales")]
         public IActionResult Create()
         {
-            var tmplist = _context.Customers.ToList();
-            ViewData["CustomerId"] = tmplist;
+            ViewData["CustomerId"] = new SelectList(_context.Customers, "CustomerId", "CustomerId");
             return View();
         }
 
         // POST: Offers/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [Authorize(Roles = "Admin,Editor,Sales")]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("NumberOfOffer,NumberOfOffer,Title,Type,Description,Author,SaleManName,DateOfCreation,LastEdition,CustomerId")] Offer offer)
+        public async Task<IActionResult> Create([Bind("OfferId,NumberOfOffer,Title,Type,Description,Author,SaleManName,DateOfCreation,Amount,TypeCurrency,LastEdition,CustomerId")] Offer offer)
         {
             if (ModelState.IsValid)
             {
@@ -69,16 +64,14 @@ namespace ProjectsControl.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            var tmplist = _context.Customers.ToList();
-            ViewData["CustomerId"] = tmplist;
+            ViewData["CustomerId"] = new SelectList(_context.Customers, "CustomerId", "CustomerId", offer.CustomerId);
             return View(offer);
         }
 
         // GET: Offers/Edit/5
-        [Authorize(Roles = "Admin,Editor,Sales")]
         public async Task<IActionResult> Edit(string id)
         {
-            if (id == null)
+            if (id == null || _context.Offers == null)
             {
                 return NotFound();
             }
@@ -88,18 +81,16 @@ namespace ProjectsControl.Controllers
             {
                 return NotFound();
             }
-            var tmplist = _context.Customers.ToList();
-            ViewData["CustomerId"] = tmplist;
+            ViewData["CustomerId"] = new SelectList(_context.Customers, "CustomerId", "CustomerId", offer.CustomerId);
             return View(offer);
         }
 
         // POST: Offers/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [Authorize(Roles = "Admin,Editor,Sales")]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(string id, [Bind("NumberOfOffer,NumberOfOffer,Title,Type,Description,Author,SaleManName,DateOfCreation,LastEdition,CustomerId")] Offer offer)
+        public async Task<IActionResult> Edit(string id, [Bind("OfferId,NumberOfOffer,Title,Type,Description,Author,SaleManName,DateOfCreation,Amount,TypeCurrency,LastEdition,CustomerId")] Offer offer)
         {
             if (id != offer.OfferId)
             {
@@ -126,16 +117,14 @@ namespace ProjectsControl.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            var tmplist = _context.Customers.ToList();
-            ViewData["CustomerId"] = tmplist;
+            ViewData["CustomerId"] = new SelectList(_context.Customers, "CustomerId", "CustomerId", offer.CustomerId);
             return View(offer);
         }
 
         // GET: Offers/Delete/5
-        [Authorize(Roles = "Admin,Editor,Sales")]
         public async Task<IActionResult> Delete(string id)
         {
-            if (id == null)
+            if (id == null || _context.Offers == null)
             {
                 return NotFound();
             }
@@ -152,20 +141,27 @@ namespace ProjectsControl.Controllers
         }
 
         // POST: Offers/Delete/5
-        [Authorize(Roles = "Admin,Editor,Sales")]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
+            if (_context.Offers == null)
+            {
+                return Problem("Entity set 'DBProjectContext.Offers'  is null.");
+            }
             var offer = await _context.Offers.FindAsync(id);
-            _context.Offers.Remove(offer);
+            if (offer != null)
+            {
+                _context.Offers.Remove(offer);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
-        [AllowAnonymous]
+
         private bool OfferExists(string id)
         {
-            return _context.Offers.Any(e => e.OfferId == id);
+          return _context.Offers.Any(e => e.OfferId == id);
         }
     }
 }
